+++ src/app.interfaces.ts
@@ -2,3 +2,10 @@ export interface IAppInfo {
 	name: string;
 	version: string;
 }
+
+export interface IMenuItem {
+	id: string;
+	label: string;
+	path?: string;
+	items?: Array<IMenuItem>;
+}