+++ src/showcase/examples/ShowcasePage.tsx
@@ -1,9 +1,0 @@
-import React, { Fragment } from 'react';
-
-export const ShowcasePage = () => {
-	return (
-		<Fragment>
-			SHOWCASE
-		</Fragment>
-	)
-}