+++ src/showcase/showcase.models.ts
@@ -1,0 +1,10 @@
+import { IMenuItem } from '../app.interfaces';
+
+// export class ShowcaseRoute {
+// 	constructor(dto: IMenuItem){
+
+// 	}
+
+// 	path: string;
+// 	// component
+// }